requirements.txt
django-dotenv
psycopg2
django==4.1.1
pytelegrambotapi
dj_database_url
qrcode==7.3.1
pillow==9.2.0
